requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
passlib[argon2]
python-jose[cryptography]
sqlalchemy
